Create Cell Geometry

Creates cell geometry from vectors of x and y positions.

create_cell_geometry(col_coords, row_coords, prj, geom = NULL,
  buffer_dist = 0)

Arguments

col_coords
numeric center positions of x indices

row_coords
numeric center positions of y indices

prj
character proj4 string for x and y

geom
sf data.frame with geometry that cell geometry should cover

buffer_dist
a distance to buffer the cell geometry in units of geom projection

Details

Intersection is performed with cell centers then geometry is constructed. A buffer may be required to fully cover geometry with cells.

Examples

library(RNetCDF)
library(sf)

#> Linking to GEOS 3.7.1, GDAL 2.3.2, PROJ 5.2.0

variable_name <- "precipitation_amount"
nc <- open.nc(system.file("extdata/metdata.nc", package = "intersectr"))
nc_file <- file.inq.nc(nc)
nc_dim <- lapply(seq(0, nc_file$ndims-1), function(d) dim.inq.nc(nc, d))
nc_var <- lapply(seq(0, nc_file$nvars-1), function(v) var.inq.nc(nc, v))
var_names <- sapply(nc_var, function(v) v$name)
print(var_names)

#> [1] "precipitation_amount" "day"                  "lat"                 
#> [4] "lon"                 

var <- which(var_names == variable_name)
var_atts <- lapply(seq(0, nc_var[[var]]$natts - 1), function(a) att.inq.nc(nc, var - 1, a))
var_att_names <- sapply(var_atts, function(a) a$name)
coordinates <- att.get.nc(nc, var - 1, which(var_att_names == "coordinates") - 1)
print(coordinates)

#> [1] "day lat lon "

# Could iomplement general handling for this... hand set for now.
# Either use coordinate variables or standard names for auxiliary coordinate variables.
x_var <- "lon"
y_var <- "lat"

if (!is.null(grid_mapping)) {
  prj <- ncdfgeom::get_prj(grid_mapping)
} else {
  prj <- "+init=epsg:4326"
  warning("No grid mapping found. Assuming WGS84")
}

#> Error in eval(expr, envir, enclos): object 'grid_mapping' not found

x <- var.get.nc(nc, x_var)
y <- var.get.nc(nc, y_var)

geom <- read_sf(system.file("shape/nc.shp", package = "sf"))

cell_geometry <- create_cell_geometry(x, y, prj, geom, 0)

#> Error in st_sf(x, ..., agr = agr, sf_column_name = sf_column_name): object 'prj' not found

plot(cell_geometry$geometry, lwd = 0.25)

#> Error in plot(cell_geometry$geometry, lwd = 0.25): object 'cell_geometry' not found

plot(st_transform(geom$geometry, prj), add = TRUE)

#> Error in make_crs(crs): object 'prj' not found